refactor: UserId 변경 Ranking, Follow 도메인에 반영

+++ src/main/java/com/moyo/backend/domain/github_ranking/data_access/RankingRepository.java
@@ -21,5 +21,5 @@ public interface RankingRepository {
 
 	Ranking findByUserId(Long userId);
 
-	Slice<Ranking> findFollowingUserRankings(List<Long> followingUserIds, RankingPeriod rankingPeriod, Pageable pageable);
+	Slice<Ranking> findFollowingUserRankings(List<Integer> followingUserIds, RankingPeriod rankingPeriod, Pageable pageable);
 }
